+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feature: add conditional page visibility
 - feature: add chart appearance controls
 - feature: choose background chart style
 - fix: hide graph entity selection for card data